Recipe functionality

+++ src/main/java/com/wildfit/server/domain/FoodItemDigest.java
@@ -13,10 +13,10 @@ public final class FoodItemDigest {
     private String foodName; //  Butter, Pure Irish, Unsalted,
     private String brandName; //  Kerrygold,
     private String brandNameItemName;
-    private Integer servingQty; //  1,
+    private Float servingQty; //  1,
     private String servingUnit; //  tbsp,
-    private Integer servingWeightGrams; //  14,
-    private Integer metricQty; //  14,
+    private Float servingWeightGrams; //  14,
+    private Float metricQty; //  14,
     private String metricUom; //  g,
     private Float calories; //  100,
     private Float totalFat; //  12,
