+++ src/main/java/com/wildfit/server/domain/IngredientDigest.java
@@ -14,34 +14,34 @@ public final class IngredientDigest {
     private long recipeId;
     private long instructionGroupId;
 
-    private String foodName; //  Butter, Pure Irish, Unsalted,
-    private String brandName; //  Kerrygold,
+    private String foodName;
+    private String brandName;
     private String brandNameItemName;
-    private Integer servingQty; //  1,
-    private String servingUnit; //  tbsp,
-    private Integer ingredientServingQty; //  1,
-    private String ingredientServingUnit; //  tbsp,
-    private Integer servingWeightGrams; //  14,
-    private Integer metricQty; //  14,
-    private String metricUom; //  g,
-    private Float calories; //  100,
-    private Float totalFat; //  12,
-    private Float saturatedFat; //  8,
-    private Float cholesterol; //  30,
-    private Float sodium; //  0,
-    private Float totalCarbohydrate; //  0,
-    private Float dietaryFiber; //  null,
-    private Float sugars; //  null,
-    private Float protein; //  0,
-    private Float potassium; //  null,
-    private Float phosphorus; //  null,
+    private Float servingQty;
+    private String servingUnit;
+    private Float ingredientServingQty;
+    private String ingredientServingUnit;
+    private Float servingWeightGrams;
+    private Float metricQty;
+    private String metricUom;
+    private Float calories;
+    private Float totalFat;
+    private Float saturatedFat;
+    private Float cholesterol;
+    private Float sodium;
+    private Float totalCarbohydrate;
+    private Float dietaryFiber;
+    private Float sugars;
+    private Float protein;
+    private Float potassium;
+    private Float phosphorus;
     private Float calcium;
     private Float iron;
     private Float vitaminD;
     private Float addedSugars;
     private Float transFattyAcid;
-    private String nixBrandName; //  Kerrygold,
-    private String nixBrandId; //  51db37b7176fe9790a8989b4,
-    private String nixItemId; //  52a15041d122497b50000a75,
+    private String nixBrandName;
+    private String nixBrandId;
+    private String nixItemId;
     private PhotoDigest photo;
 }
