+++ src/main/java/com/wildfit/server/model/SearchedFoodItem.java
@@ -9,9 +9,9 @@ public class SearchedFoodItem {
     private String food_name; //  Butter, Pure Irish, Unsalted,
     private String brand_name; //  Kerrygold,
     private String brand_name_item_name; // Kerrygold Butter Sticks, Pure Irish, Unsalted
-    private Integer serving_qty; //  1,
+    private Float serving_qty; //  1,
     private String serving_unit; //  tbsp,
-    private Integer nf_calories; //  100,
+    private Float nf_calories; //  100,
     private String nix_brand_name; //  Kerrygold,
     private String nix_brand_id; //  51db37b7176fe9790a8989b4,
     private String nix_item_id; //  52a15041d122497b50000a75,
